pyproject.toml
[project]
name = "uv-fastapi-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "fastapi[standard]>=0.112.1",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.41",
    "sqlmodel>=0.0.27",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "sqlacodegen>=3.1.1",
]
